+++ inlang/source-code/sdk/src/loadProject.ts
@@ -42,6 +42,8 @@ import { createNodeishFsWithWatcher } from "./createNodeishFsWithWatcher.js"
 
 import { maybeCreateFirstProjectId } from "./migrations/maybeCreateFirstProjectId.js"
 
+import { capture } from "./telemetry/capture.js"
+
 const settingsCompiler = TypeCompiler.Compile(ProjectSettings)
 
 /**
@@ -51,14 +53,22 @@ const settingsCompiler = TypeCompiler.Compile(ProjectSettings)
  * @param @deprecated nodeishFs - Filesystem that implements the NodeishFilesystemSubset interface.
  * @param _import - Use `_import` to pass a custom import function for testing,
  *   and supporting legacy resolvedModules such as CJS.
- * @param _capture - Use `_capture` to capture events for analytics.
  *
  */
 export async function loadProject(args: {
 	projectPath: string
 	nodeishFs: Repository["nodeishFs"]
+	/**
+	 * The app id is used to identify the app that is using the SDK.
+	 *
+	 * We use the app id to group events in telemetry to answer questions
+	 * like "Which apps causes these errors?" or "Which apps are used more than others?".
+	 *
+	 * @example
+	 * 	appId: "app.inlang.badge"
+	 */
+	appId?: string
 	_import?: ImportFunction
-	_capture?: (id: string, props: Record<string, unknown>) => void
 }): Promise<InlangProject>
 
 /**
@@ -66,21 +76,20 @@ export async function loadProject(args: {
  * @param repo - An instance of a lix repo as returned by `openRepository`.
  * @param _import - Use `_import` to pass a custom import function for testing,
  *   and supporting legacy resolvedModules such as CJS.
- * @param _capture - Use `_capture` to capture events for analytics.
  *
  */
 export async function loadProject(args: {
 	projectPath: string
 	repo: Repository
+	appId?: string
 	_import?: ImportFunction
-	_capture?: (id: string, props: Record<string, unknown>) => void
 }): Promise<InlangProject>
 
 export async function loadProject(args: {
 	projectPath: string
 	repo?: Repository
+	appId?: string
 	_import?: ImportFunction
-	_capture?: (id: string, props: Record<string, unknown>) => void
 	nodeishFs?: Repository["nodeishFs"]
 }): Promise<InlangProject> {
 	const projectPath = normalizePath(args.projectPath)
@@ -123,7 +132,15 @@ export async function loadProject(args: {
 	await maybeCreateFirstProjectId({ projectPath, repo: args.repo })
 
 	// -- load project ------------------------------------------------------
+
 	return await createRoot(async () => {
+		// TODO remove tryCatch after https://github.com/opral/monorepo/issues/2013
+		// - a repo will always be present
+		// - if a repo is present, the project id will always be present
+		const { data: projectId } = await tryCatch(() =>
+			fs.readFile(args.projectPath + "/project_id", { encoding: "utf-8" })
+		)
+
 		const [initialized, markInitAsComplete, markInitAsFailed] = createAwaitable()
 		// -- settings ------------------------------------------------------------
 
@@ -137,12 +154,7 @@ export async function loadProject(args: {
 			// }
 
 			loadSettings({ settingsFilePath: projectPath + "/settings.json", nodeishFs })
-				.then((settings) => {
-					setSettings(settings)
-					// rename settings to get a convenient access to the data in Posthog
-					const project_settings = settings
-					args._capture?.("SDK used settings", { project_settings })
-				})
+				.then((settings) => setSettings(settings))
 				.catch((err) => {
 					markInitAsFailed(err)
 				})
@@ -588,6 +600,29 @@ export async function loadProject(args: {
 		// 		)
 		// 	)
 
+		/**
+		 * Utility to escape reactive tracking and avoid multiple calls to
+		 * the capture event.
+		 *
+		 * Should be addressed with https://github.com/opral/monorepo/issues/1772
+		 */
+		let projectLoadedCapturedAlready = false
+
+		if (projectId && projectLoadedCapturedAlready === false) {
+			projectLoadedCapturedAlready = true
+			// TODO ensure that capture is "awaited" without blocking the the app from starting
+			await capture("SDK loaded project", {
+				projectId,
+				properties: {
+					appId: args.appId,
+					settings: settings(),
+					installedPluginIds: installedPlugins().map((p) => p.id),
+					installedMessageLintRuleIds: installedMessageLintRules().map((r) => r.id),
+					numberOfMessages: messagesQuery.includedMessageIds().length,
+				},
+			})
+		}
+
 		return {
 			installed: {
 				plugins: createSubscribable(() => installedPlugins()),
@@ -775,7 +810,7 @@ async function loadMessages(
 	fs: NodeishFilesystemSubset,
 	messagesQuery: InlangProject["query"]["messages"],
 	settingsValue: ProjectSettings,
-	loadPlugin: Plugin["loadMessages"]
+	loadPlugin: any
 ) {
 	// the current approach introuces a sync between both systems - the legacy load / save messages plugins and the new format - we dont delete messages that we don't see int he plugins produced messages array anymore
 
@@ -809,7 +844,7 @@ async function loadMessages(
 		const currentMessages = messagesQuery
 			.getAll()
 			// TODO #1585 here we match using the id to support legacy load message plugins - after we introduced import / export methods we will use importedMessage.alias
-			.filter((message: any) => message.alias[loadPluginId] === loadedMessage.id)
+			.filter((message: any) => message.alias["default"] === loadedMessage.id)
 
 		if (currentMessages.length > 1) {
 			// NOTE: if we happen to find two messages witht the sam alias we throw for now
@@ -819,8 +854,7 @@ async function loadMessages(
 			// update message in place - leave message id and alias untouched
 			loadedMessage.alias = {} as any
 			// TODO #1585 we have to map the id of the importedMessage to the alias and fill the id property with the id of the existing message - change when import mesage provides importedMessage.alias
-			loadedMessage.alias[loadPluginId] = loadedMessage.id
-			loadedMessage.alias["library.inlang.paraglideJs"] = loadedMessage.id
+			loadedMessage.alias["default"] = loadedMessage.id
 			loadedMessage.id = currentMessages[0]!.id
 
 			// TODO #1844 INFORM stringifyMessage encodes messages independent from key order!
@@ -834,8 +868,7 @@ async function loadMessages(
 			// message with the given alias does not exist so far
 			loadedMessage.alias = {} as any
 			// TODO #1585 we have to map the id of the importedMessage to the alias - change when import mesage provides importedMessage.alias
-			loadedMessage.alias[loadPluginId] = loadedMessage.id
-			loadedMessage.alias["library.inlang.paraglideJs"] = loadedMessage.id
+			loadedMessage.alias["default"] = loadedMessage.id
 
 			let currentOffset = 0
 			let messsageId: string | undefined
@@ -880,7 +913,7 @@ async function saveMessages(
 	fs: NodeishFilesystemSubset,
 	messagesQuery: InlangProject["query"]["messages"],
 	settingsValue: ProjectSettings,
-	savePlugin: Plugin["saveMessages"]
+	savePlugin: any
 ) {
 	// queue next save if we have a save ongoing
 	if (isSaving) {
@@ -904,7 +937,7 @@ async function saveMessages(
 			for (const message of currentMessages) {
 				const fixedExportMessage = { ...message }
 				// TODO #1585 here we match using the id to support legacy load message plugins - after we introduced import / export methods we will use importedMessage.alias
-				fixedExportMessage.id = fixedExportMessage.alias[savePluginId] ?? fixedExportMessage.id
+				fixedExportMessage.id = fixedExportMessage.alias["default"] ?? fixedExportMessage.id
 
 				messagesToExport.push(fixedExportMessage)
 			}