+++ inlang/source-code/sdk/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @inlang/sdk
 
+## 0.23.0
+
+### Minor Changes
+
+- b920761e6: refactor: remove `_capture` in favor of `appDir` for product insights
+
+### Patch Changes
+
+- Updated dependencies [74ac11c47]
+  - @inlang/language-tag@1.4.0
+  - @inlang/message@2.0.1
+  - @inlang/message-lint-rule@1.4.1
+  - @inlang/plugin@2.4.2
+  - @inlang/project-settings@2.2.1
+  - @inlang/translatable@1.2.1
+  - @inlang/module@1.2.2
+
 ## 0.22.0
 
 ### Minor Changes