+++ inlang/source-code/versioned-interfaces/project-settings/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @inlang/project-settings
 
+## 2.2.1
+
+### Patch Changes
+
+- Updated dependencies [74ac11c47]
+  - @inlang/language-tag@1.4.0
+
 ## 2.2.0
 
 ### Minor Changes